feat: add CRUD functionality for social profile

+++ src/app/services/socialProfile.service.ts
@@ -1,26 +1,69 @@
+import { ICreateSocialProfileRequest } from '@/app/interfaces/requests';
+import SocialProfile from '@/app/models/socialProfile.model';
+import { AppError } from '@/app/utils';
+import { Types } from 'mongoose';
+// import { Schema } from 'mongoose';
+
 class SocialProfileService {
-    create = async (data: any): Promise<any> => {
+    create = async (data: ICreateSocialProfileRequest): Promise<any> => {
         // Implement create logic
-        return {};
+        const res = await SocialProfile.create({ ...data, userId: data.userId });
+        return res;
     };
 
-    getAll = async (): Promise<any[]> => {
+    getAll = async (userId: Types.ObjectId): Promise<any[]> => {
         // Implement get all logic
-        return [];
+        return await SocialProfile.find({ userId });
     };
 
     getById = async (id: string): Promise<any> => {
         // Implement get by id logic
-        return {};
+        const result = await SocialProfile.findById(id);
+
+        if (!result) {
+            throw new AppError('Social profile not found', 404);
+        }
+
+        return;
     };
 
-    updateById = async (id: string, data: any): Promise<any> => {
+    updateById = async (
+        id: string,
+        userId: Types.ObjectId,
+        data: Partial<ICreateSocialProfileRequest>
+    ): Promise<any> => {
         // Implement update by id logic
-        return {};
+
+        // check existence
+        const profile = await SocialProfile.findById(id);
+        if (!profile) {
+            throw new AppError('Social profile not found', 404);
+        }
+
+        // check ownership
+        if (profile.userId.toString() !== userId.toString()) {
+            throw new AppError('You do not have permission to update this social profile', 403);
+        }
+
+        return await SocialProfile.findByIdAndUpdate(id, data, { new: true, runValidators: true });
     };
 
-    deleteById = async (id: string): Promise<void> => {
+    deleteById = async (id: string, userId: Types.ObjectId): Promise<void> => {
         // Implement delete by id logic
+
+        // check existence
+        const profile = await SocialProfile.findById(id);
+        if (!profile) {
+            throw new AppError('Social profile not found', 404);
+        }
+
+        // check ownership
+        if (profile.userId.toString() !== userId.toString()) {
+            throw new AppError('You do not have permission to update this social profile', 403);
+        }
+
+        await SocialProfile.findByIdAndDelete(id);
+        return;
     };
 }
